% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{tree}
\alias{tree}
\alias{create}
\alias{addnode}
\alias{value}
\title{Create a decision tree}
\usage{
tree(x)

create(x, name, type, n, sub_name, prob = NA)

addnode(x, name, type = NA, n = NA, sub_name = NA, prob = NA,
  leaf_nodevalue = NA)

value(x, opti = "max")
}
\arguments{
\item{x}{an object of the tree}

\item{name}{name of a decision node}

\item{type}{character, means type of a decision node (chance or choice)}

\item{n}{number of the branches of a node}

\item{sub_name}{vector, the names of each branch, cannot be repeated}

\item{prob}{vector, the probability for each branch. sum should be 1}

\item{opti}{character, "max" or "min" for type choice}

\item{leafnode_value}{vector, the value of the leaf node}
}
\description{
the list create stores all the info needed for a decision tree structure
function tree() for initializing teh tree
function create() for adding root node
function addnode() for adding rest nodes
function eval() for calculate the value for each node
}
\examples{

k <- tree()
k <- create(k,"a","choice",2,c("treasury","LLC"))
k <- addnode(k,"LLC","chance",2,c("fav1","unfav1"),c(0.3,0.7))
k <- addnode(k,"fav1","chance",2,c("yes","no"),prob=c(0.5,0.5),leaf_nodevalue=list(7000,20000))
k <- addnode(k,"unfav1",leaf_nodevalue=list(190000))
k <- addnode(k,"treasury","chance",2,c("unfav2","fav2"),c(0.75,0.25))
k <- addnode(k,"fav2","chance",2,prob=c(0.75,0.25),leaf_nodevalue=list(-110000,190000))
k <- addnode(k,"unfav2",leaf_nodevalue=list(2000))
k <- value(k,"max")

# we can input some random numbers as leaf node value
k <- tree()
k <- create(k,"a","choice",2,c("treasury","LLC"))
k <- addnode(k,"LLC","chance",2,c("fav1","unfav1"),c(0.3,0.7))
k <- addnode(k,"fav1","chance",2,c("yes","no"),prob=c(0.5,0.5),leaf_nodevalue=list(rnorm(10000,7000,100),20000))
k <- addnode(k,"unfav1",leaf_nodevalue=list(190000))
k <- addnode(k,"treasury","chance",2,c("unfav2","fav2"),c(0.75,0.25))
k <- addnode(k,"fav2","chance",2,prob=c(0.75,0.25),leaf_nodevalue=list(rnorm(100,-110000,20000),190000))
k <- addnode(k,"unfav2",leaf_nodevalue=list(2000))
k <- value(k,"max")

}
\author{
Kevin Hu \email{kevinhu@bu.edu}
}
