% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{uniform}
\alias{uniform}
\alias{normal}
\alias{gamma}
\alias{triangular}
\alias{cumul}
\alias{meanci}
\alias{target}
\alias{citrials}
\title{generate random numbers}
\usage{
uniform(n, min, max)

normal(n, mean, stdev)

gamma(n, shape, scale)

triangular(n, min, likely, max)

cumul(n, min, max, range, prob)

meanci(mean, stdev, n, alpha = 0.95)

target(data, value)

citrials(stdev, width, alpha = 0.95)
}
\arguments{
\item{n}{number of observations}

\item{min}{the lower bound of a distribution}

\item{max}{the higher bound of a diatribution}

\item{mean}{vector of means}

\item{stdev}{vector of standard deviations}

\item{shape, scale}{shape and scale parameters. Must be positive, scale strictly}

\item{likely}{the most likely number for triangular diatribution}

\item{range}{a vector for breaks of a distribution}

\item{prob}{cumulative probability for each breaks}

\item{alpha}{confidence degree}

\item{width}{required width for confidence interval}
}
\description{
distribution helps generate random numbers with different diatributions
It includes uniform, normal, gamma, triangular, cumulative distributions.
confidence interval of a mean value, percentile of a number among a
list of numbers and the least number of trials for a specific width of confidence
interval are included
}
\examples{
a <- triangular(10000,100,130,200)
b <- cumul(1000,3,10,c(4,6),c(0.5,0.6))
meanci(100,20,1000,alpha=0.9)
target(b,0.8)
citrials(20,1,alpha=0.9)


}
\author{
Kevin Hu \email{kevinhu@bu.edu}
}
