% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{ks_fitdist}
\alias{ks_fitdist}
\title{fit a distribution}
\usage{
ks_fitdist(x)
}
\arguments{
\item{x}{vector, numbers for fitting a distribution}
}
\description{
Given a bunch of numbers, ks_fitdist() gives the distribution best fitting
the numbers. fitdistr() in MASS package is used to obtain the parameters. The
rank of the distributions is based on Kolmogorov-Smirnov Tests. Distributions
include beta, cauthy, chi-square, exponential, f, gamma, geometric, lognormal,
logistic, negative binomial, normal, Poisson, t, weibull. Histograms for comparison
are shown.
}
\examples{
b <- rgamma(10000,20,1)
ks_fitdist(b)

}
\author{
Kevin Hu \email{kevinhu@bu.edu}
}
